teamenigma-master/requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
contextlib2==0.6.0.post1
dateparser==1.0.0
decorator==5.0.5
defusedxml==0.7.1
distlib==0.3.1
docutils==0.16
elasticsearch==7.12.0
pyOpenSSL==20.0.1
python-nmap==0.6.4
python-whois==0.7.3
python3-nmap==1.5.0
pytz==2021.1
regex==2021.4.4
requests==2.26.0
schema==0.7.4
selenium==3.141.0
simplejson==3.17.3
six==1.15.0
urllib3==1.26.4
python3-nmap


